Move edit button from EntityActivity to EntityData
