+++ src/components/entity/activity.vue
@@ -13,10 +13,6 @@ except according to the terms contained in the LICENSE file.
   <page-section id="entity-activity">
     <template #heading>
       <span>{{ $t('common.activity') }}</span>
-      <button v-if="rendersUpdateButton" id="entity-activity-update-button"
-        type="button" class="btn btn-default" @click="$emit('update')">
-        <span class="icon-pencil"></span>{{ $t('action.edit') }}
-      </button>
     </template>
     <template #body>
       <loading :state="initiallyLoading"/>
@@ -42,15 +38,10 @@ import PageSection from '../page/section.vue';
 
 import { useRequestData } from '../../request-data';
 
-defineEmits(['update']);
-
 // The component does not assume that this data will exist when the component is
 // created.
-const { project, dataset, audits, diffs, resourceStates } = useRequestData();
+const { audits, diffs, resourceStates } = useRequestData();
 const { initiallyLoading, dataExists } = resourceStates([audits, diffs]);
-
-const rendersUpdateButton = computed(() => project.dataExists &&
-  project.permits('entity.update') && dataset.dataExists);
 
 const feed = computed(() => {
   const result = [];
